Module Posix_signal

POSIX signal handling bindings.

This module provides OCaml bindings to POSIX signal functions defined in signal.h.

It includes functions for manipulating signal sets and controlling signal delivery to threads and processes.

Signal Sets

type sigset

Abstract type representing a set of signals.

type signal = [

| `Sigabrt
(*
Abnormal termination
*)

| `Sigalrm
(*
Alarm clock
*)

| `Sigbus
(*
Bus error
*)

| `Sigchld
(*
Child process terminated
*)

| `Sigcont
(*
Continue if stopped
*)

| `Sigfpe
(*
Floating-point exception
*)

| `Sighup
(*
Hangup
*)

| `Sigill
(*
Illegal instruction
*)

| `Sigint
(*
Interactive attention signal
*)

| `Sigkill
(*
Kill (cannot be caught or ignored)
*)

| `Sigpipe
(*
Broken pipe
*)

| `Sigquit
(*
Quit
*)

| `Sigsegv
(*
Segmentation fault
*)

| `Sigstop
(*
Stop (cannot be caught or ignored)
*)

| `Sigterm
(*
Termination request
*)

| `Sigtstp
(*
Terminal stop
*)

| `Sigttin
(*
Background read from terminal
*)

| `Sigttou
(*
Background write to terminal
*)

| `Sigusr1
(*
User-defined signal 1
*)

| `Sigusr2
(*
User-defined signal 2
*)

| `Sigtrap
(*
Trace/breakpoint trap
*)

| `Sigurg
(*
Urgent data on socket
*)

| `Sigxcpu
(*
CPU time limit exceeded
*)

| `Sigxfsz

]

POSIX signals.

File size limit exceeded

Signal Mask Actions

type action = [

| `Sig_block
(*
Add signals to the mask
*)

| `Sig_setmask
(*
Set the mask to the specified signals
*)

| `Sig_unblock
(*
Remove signals from the mask
*)

]

Actions for modifying the signal mask.

Signal Set Functions

val sigemptyset : unit -> sigset

Create an empty signal set. See sigemptyset(3).

val sigaddset : sigset -> signal -> unit

Add a signal to a signal set. See sigaddset(3).

val sigismember : sigset -> signal -> bool

Test whether a signal is in a signal set. See sigismember(3).

Signal Mask Functions

val pthread_sigmask : action -> sigset option -> sigset

Examine and change the signal mask of the calling thread.

See pthread_sigmask(3).

parameter action
How to modify the signal mask.

parameter set
The signal set to use, or None to only retrieve the current mask.

returns
The previous signal mask.

val sigprocmask : action -> sigset option -> sigset

Examine and change the signal mask of the calling process.

See sigprocmask(2).

Note: In multi-threaded programs, use pthread_sigmask instead.

parameter action
How to modify the signal mask.

parameter set
The signal set to use, or None to only retrieve the current mask.

returns
The previous signal mask.